feat(app): centralize map location resolver and region labels (#1347)
- Add map_location_resolver.dart (province/sea tile keys for locate-on-map).
- Add region_labels.dart (regionDisplayLabel); unitsPanelRegionLabel delegates.
- Wire naval/military panels and split_fleet_dialog; military tests import resolver.

Closes #1347.

Made-with: Cursor
