+++ app/lib/features/game/widgets/units/shared/units_panel_region_label.dart
@@ -1,13 +1,6 @@
 // Shared region labels for units panels. SPEC/ui/civilian-units-panel.md et al.
 
+import '../../../utils/region_labels.dart';
+
 /// Region id to display label for Old/New World sections in units panels.
-String unitsPanelRegionLabel(String regionId) {
-  switch (regionId) {
-    case 'oldWorld':
-      return 'Old World';
-    case 'newWorld':
-      return 'New World';
-    default:
-      return regionId;
-  }
-}
+String unitsPanelRegionLabel(String regionId) => regionDisplayLabel(regionId);
